Move PluginsMinVersions and ExtensionUtils in plugins-common
