+++ core/src/main/java/org/sonarsource/sonarlint/core/container/standalone/StandaloneGlobalContainer.java
@@ -44,7 +44,6 @@ import org.sonarsource.sonarlint.core.client.api.standalone.StandaloneRuleDetail
 import org.sonarsource.sonarlint.core.commons.progress.ProgressMonitor;
 import org.sonarsource.sonarlint.core.container.ComponentContainer;
 import org.sonarsource.sonarlint.core.container.analysis.AnalysisContainer;
-import org.sonarsource.sonarlint.core.container.connected.validate.PluginVersionChecker;
 import org.sonarsource.sonarlint.core.container.global.ExtensionInstaller;
 import org.sonarsource.sonarlint.core.container.global.GlobalConfigurationProvider;
 import org.sonarsource.sonarlint.core.container.global.GlobalExtensionContainer;
@@ -62,6 +61,7 @@ import org.sonarsource.sonarlint.core.plugin.PluginInstancesLoader;
 import org.sonarsource.sonarlint.core.plugin.PluginRepository;
 import org.sonarsource.sonarlint.core.plugin.cache.PluginCacheProvider;
 import org.sonarsource.sonarlint.core.plugin.commons.ApiVersions;
+import org.sonarsource.sonarlint.core.plugin.commons.PluginsMinVersions;
 
 public class StandaloneGlobalContainer extends ComponentContainer {
 
@@ -85,7 +85,7 @@ public class StandaloneGlobalContainer extends ComponentContainer {
       new StandalonePluginUrls(globalConfig.getPluginUrls()),
       StandalonePluginIndex.class,
       PluginRepository.class,
-      PluginVersionChecker.class,
+      PluginsMinVersions.class,
       PluginInfosLoader.class,
       PluginInstancesLoader.class,
       PluginClassloaderFactory.class,
